# Support 'NULL' as Null in csv parser.