--- a/datafusion/proto-common/src/generated/prost.rs
+++ b/datafusion/proto-common/src/generated/prost.rs
@@ -593,17 +593,20 @@ pub struct CsvOptions {
     /// Optional representation of null value
     #[prost(string, tag = "12")]
     pub null_value: ::prost::alloc::string::String,
+    /// Optional representation of null loading regex
+    #[prost(string, tag = "13")]
+    pub null_regex: ::prost::alloc::string::String,
     /// Optional comment character as a byte
-    #[prost(bytes = "vec", tag = "13")]
+    #[prost(bytes = "vec", tag = "14")]
     pub comment: ::prost::alloc::vec::Vec<u8>,
     /// Indicates if quotes are doubled
-    #[prost(bytes = "vec", tag = "14")]
+    #[prost(bytes = "vec", tag = "15")]
     pub double_quote: ::prost::alloc::vec::Vec<u8>,
     /// Indicates if newlines are supported in values
-    #[prost(bytes = "vec", tag = "15")]
+    #[prost(bytes = "vec", tag = "16")]
     pub newlines_in_values: ::prost::alloc::vec::Vec<u8>,
     /// Optional terminator character as a byte
-    #[prost(bytes = "vec", tag = "16")]
+    #[prost(bytes = "vec", tag = "17")]
     pub terminator: ::prost::alloc::vec::Vec<u8>,
 }
 /// Options controlling CSV format
@@ -638,27 +641,33 @@ pub struct ParquetColumnSpecificOptions {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ParquetColumnOptions {
     #[prost(oneof = "parquet_column_options::BloomFilterEnabledOpt", tags = "1")]
-    pub bloom_filter_enabled_opt:
-        ::core::option::Option<parquet_column_options::BloomFilterEnabledOpt>,
+    pub bloom_filter_enabled_opt: ::core::option::Option<
+        parquet_column_options::BloomFilterEnabledOpt,
+    >,
     #[prost(oneof = "parquet_column_options::EncodingOpt", tags = "2")]
     pub encoding_opt: ::core::option::Option<parquet_column_options::EncodingOpt>,
     #[prost(oneof = "parquet_column_options::DictionaryEnabledOpt", tags = "3")]
-    pub dictionary_enabled_opt:
-        ::core::option::Option<parquet_column_options::DictionaryEnabledOpt>,
+    pub dictionary_enabled_opt: ::core::option::Option<
+        parquet_column_options::DictionaryEnabledOpt,
+    >,
     #[prost(oneof = "parquet_column_options::CompressionOpt", tags = "4")]
     pub compression_opt: ::core::option::Option<parquet_column_options::CompressionOpt>,
     #[prost(oneof = "parquet_column_options::StatisticsEnabledOpt", tags = "5")]
-    pub statistics_enabled_opt:
-        ::core::option::Option<parquet_column_options::StatisticsEnabledOpt>,
+    pub statistics_enabled_opt: ::core::option::Option<
+        parquet_column_options::StatisticsEnabledOpt,
+    >,
     #[prost(oneof = "parquet_column_options::BloomFilterFppOpt", tags = "6")]
-    pub bloom_filter_fpp_opt:
-        ::core::option::Option<parquet_column_options::BloomFilterFppOpt>,
+    pub bloom_filter_fpp_opt: ::core::option::Option<
+        parquet_column_options::BloomFilterFppOpt,
+    >,
     #[prost(oneof = "parquet_column_options::BloomFilterNdvOpt", tags = "7")]
-    pub bloom_filter_ndv_opt:
-        ::core::option::Option<parquet_column_options::BloomFilterNdvOpt>,
+    pub bloom_filter_ndv_opt: ::core::option::Option<
+        parquet_column_options::BloomFilterNdvOpt,
+    >,
     #[prost(oneof = "parquet_column_options::MaxStatisticsSizeOpt", tags = "8")]
-    pub max_statistics_size_opt:
-        ::core::option::Option<parquet_column_options::MaxStatisticsSizeOpt>,
+    pub max_statistics_size_opt: ::core::option::Option<
+        parquet_column_options::MaxStatisticsSizeOpt,
+    >,
 }
 /// Nested message and enum types in `ParquetColumnOptions`.
 pub mod parquet_column_options {
@@ -763,22 +772,27 @@ pub struct ParquetOptions {
     #[prost(string, tag = "16")]
     pub created_by: ::prost::alloc::string::String,
     #[prost(oneof = "parquet_options::MetadataSizeHintOpt", tags = "4")]
-    pub metadata_size_hint_opt:
-        ::core::option::Option<parquet_options::MetadataSizeHintOpt>,
+    pub metadata_size_hint_opt: ::core::option::Option<
+        parquet_options::MetadataSizeHintOpt,
+    >,
     #[prost(oneof = "parquet_options::CompressionOpt", tags = "10")]
     pub compression_opt: ::core::option::Option<parquet_options::CompressionOpt>,
     #[prost(oneof = "parquet_options::DictionaryEnabledOpt", tags = "11")]
-    pub dictionary_enabled_opt:
-        ::core::option::Option<parquet_options::DictionaryEnabledOpt>,
+    pub dictionary_enabled_opt: ::core::option::Option<
+        parquet_options::DictionaryEnabledOpt,
+    >,
     #[prost(oneof = "parquet_options::StatisticsEnabledOpt", tags = "13")]
-    pub statistics_enabled_opt:
-        ::core::option::Option<parquet_options::StatisticsEnabledOpt>,
+    pub statistics_enabled_opt: ::core::option::Option<
+        parquet_options::StatisticsEnabledOpt,
+    >,
     #[prost(oneof = "parquet_options::MaxStatisticsSizeOpt", tags = "14")]
-    pub max_statistics_size_opt:
-        ::core::option::Option<parquet_options::MaxStatisticsSizeOpt>,
+    pub max_statistics_size_opt: ::core::option::Option<
+        parquet_options::MaxStatisticsSizeOpt,
+    >,
     #[prost(oneof = "parquet_options::ColumnIndexTruncateLengthOpt", tags = "17")]
-    pub column_index_truncate_length_opt:
-        ::core::option::Option<parquet_options::ColumnIndexTruncateLengthOpt>,
+    pub column_index_truncate_length_opt: ::core::option::Option<
+        parquet_options::ColumnIndexTruncateLengthOpt,
+    >,
     #[prost(oneof = "parquet_options::EncodingOpt", tags = "19")]
     pub encoding_opt: ::core::option::Option<parquet_options::EncodingOpt>,
     #[prost(oneof = "parquet_options::BloomFilterFppOpt", tags = "21")]
@@ -861,9 +875,7 @@ pub struct ColumnStats {
     #[prost(message, optional, tag = "4")]
     pub distinct_count: ::core::option::Option<Precision>,
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum JoinType {
     Inner = 0,
@@ -910,9 +922,7 @@ impl JoinType {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum JoinConstraint {
     On = 0,
@@ -938,9 +948,7 @@ impl JoinConstraint {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum TimeUnit {
     Second = 0,
@@ -972,9 +980,7 @@ impl TimeUnit {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum IntervalUnit {
     YearMonth = 0,
@@ -1003,9 +1009,7 @@ impl IntervalUnit {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum UnionMode {
     Sparse = 0,
@@ -1031,9 +1035,7 @@ impl UnionMode {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum CompressionTypeVariant {
     Gzip = 0,
@@ -1068,9 +1070,7 @@ impl CompressionTypeVariant {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum JoinSide {
     LeftSide = 0,
@@ -1099,9 +1099,7 @@ impl JoinSide {
         }
     }
 }
-#[derive(
-    Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration,
-)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum PrecisionInfo {
     Exact = 0,