… datafusion-cli/Cargo.lock
… datafusion/core/tests/data/csv/aggregate_test_100_with_nulls.csv