@jayzhan211 2024-11-02
I think we can also add such test under `datafusion/core/tests/data/`