@dhegberg 2024-11-02
@alamb These tests will fail until the associated data changes are merged: https://github.com/apache/arrow-testing/pull/103

Let me know if there is a preferred way of setting up tests for these changes.