@dhegberg 2024-12-17
@jayzhan211 @alamb 

I've revised this to set the null import via config.

Shows a small regression when this config is used, but it should be acceptable since this is opt in.